- Directories restructured
* main.cpp was removed along with the related files
* An examples folder was made to add future executables
* An example template folder was added to the examples directory
