+++ lib/GDMBasics/CMakeLists.txt
@@ -7,12 +7,12 @@ project(
         LANGUAGES CXX
 )
 
-# Header files
-include_directories(include)
-
 # Build library
-file(GLOB_RECURSE GDM_SOURCE_CODE "src/*cpp")
+file(GLOB_RECURSE GDM_SOURCE_CODE ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
 add_library(${PROJECT_NAME} STATIC ${GDM_SOURCE_CODE})
+
+# Header files
+target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
 
 target_link_libraries(GDMBasics
         sfml-graphics
